+++ DDHidLib/DDHidKeyboardBarcodeScanner.m
@@ -95,7 +95,7 @@
     return mKeyElements;
 }
 
-- (unsigned) numberOfKeys;
+- (NSUInteger) numberOfKeys;
 {
     return [mKeyElements count];
 }
